Settle merge conflicts

+++ ext/share/botUtils.py
@@ -27,10 +27,20 @@ def subPerms(ctx):
     return userPerms.administrator or userPerms.manage_webhooks or ctx.guild.owner_id == ctx.author.id
 
 async def msgDelete(ctx: Union[commands.Context, SlashContext]):
+    """
+    Removes the message that invoked the command (if any)
+
+    Arguments:
+    ---
+    `ctx`: A discord.py `commands.Context` or discord-py-slash-commands `SlashContext`
+    """
     if type(ctx) != SlashContext:
         await ctx.message.delete()
 
 class fandomTextParse():
+    """
+    Class that contains functions for Fandom Wiki related actions.
+    """
     async def parseToEmbed(name: str, embedData: list):
         embedParts = {}
         excessParts = None
@@ -68,7 +78,8 @@ class fandomTextParse():
             elif headerCount > 1:
                 partPhrase = "these sections have"
             discordEmbed.add_field(name=fieldHeader.strip("/"), value=f"Due to Discord's limitations, {partPhrase} to be seperated to different embeds.\n"
-                                                                      f"Respond with {partChoice.strip(',')} to look at the section.")
+                                                                      f"Respond with {partChoice.strip(',')} to look at the section.", inline=False)
+            discordEmbed.add_field(name="Not the VTuber you're searching for?", value=f"Respond with `search` to choose the correct VTuber.", inline=False)
 
         discordEmbed.set_footer(text="Powered by Fandom", icon_url="https://img.ezz.moe/0407/14-24-14.png")
         return discordEmbed, excessParts
@@ -104,48 +115,65 @@ class fandomTextParse():
 
         return dictText
 
-async def vtuberSearch(ctx: Union[commands.Context, SlashContext], bot: commands.Bot, searchTerm: str, searchMsg):
+async def vtuberSearch(ctx: Union[commands.Context, SlashContext], bot: commands.Bot, searchTerm: str, searchMsg, askTerm: str, getOther: bool = False):
+    """
+    Searches for a VTuber and returns a `dict` with it's relevant data.
+
+    Arguments:
+    ---
+    `ctx`: A `discord.py` command context or a `discord-py-slash-command` slash context.
+    `bot`: A `discord.py` `commands.Bot` object.
+    `searchTerm`: The search term that is passed by from the user's input.
+    `searchMsg`: A Discord message object that is responsible for the search message.
+    `askTerm`: Term used when the search embed says "[askTerm] this channel"
+    `getOther`: Whether to have a message confirming the detected VTuber from `searchTerm` or to assume that the first term is correct.
+    """
     from ..infoscraper import FandomScrape, channelInfo
+
     getChannel = False
 
-    if "https://www.youtube.com/channel/" in searchTerm:
-        for part in searchTerm.split("/"):
-            if 23 <= len(part) <= 25:
-                if part[0] == "U":
-                    channelID = part
-                    getChannel = True
-    
-    if 23 <= len(searchTerm) <= 25:
-        if searchTerm[0] == "U":
-            async with aiohttp.ClientSession() as session:
-                async with session.get(f"https://www.youtube.com/channel/{searchTerm}") as r:
-                    if r.status == 200:
-                        cInfo = await channelInfo(channelID)
-                        if not cInfo["success"]:
-                            cInfo = await channelInfo(channelID, True)
-                        sConfirm = await searchConfirm(ctx, bot, cInfo["name"], searchMsg, f"Subscribe to {cInfo['name']}?", "Subscribe to this channel", "Choose another channel")
-                        if sConfirm["success"]:
-                            return {
-                                "success": True,
-                                "channelID": searchTerm,
-                                "name": cInfo['name']
-                            }
-                        elif not sConfirm["success"] and not sConfirm["declined"]:
-                            await searchTerm.delete()
-                            await msgDelete(ctx)
-                            return {
-                                "success": False
-                            }
+    if not getOther:
+        if "https://www.youtube.com/channel/" in searchTerm:
+            for part in searchTerm.split("/"):
+                if 23 <= len(part) <= 25:
+                    if part[0] == "U":
+                        channelID = part
+                        getChannel = True
+        
+        if 23 <= len(searchTerm) <= 25:
+            if searchTerm[0] == "U":
+                async with aiohttp.ClientSession() as session:
+                    async with session.get(f"https://www.youtube.com/channel/{searchTerm}") as r:
+                        if r.status == 200:
+                            cInfo = await channelInfo(searchTerm)
+                            if not cInfo["success"]:
+                                cInfo = await channelInfo(searchTerm, True)
+                            sConfirm = await searchConfirm(ctx, bot, cInfo["name"], searchMsg, f"{askTerm} {cInfo['name']}?", f"{askTerm} this channel", "Choose another channel", True)
+                            if sConfirm["success"]:
+                                return {
+                                    "success": True,
+                                    "channelID": searchTerm,
+                                    "name": cInfo['name']
+                                }
+                            elif not sConfirm["success"] and not sConfirm["declined"]:
+                                await searchMsg.delete()
+                                await msgDelete(ctx)
+                                return {
+                                    "success": False
+                                }
     
     if not getChannel:
         fandomSearch = await FandomScrape.searchChannel(searchTerm)
 
-        if fandomSearch["status"] == "Success":
-            sConfirm = await searchConfirm(ctx, bot, fandomSearch["name"], searchMsg, f"Subscribe to {fandomSearch['name']}?", "Subscribe to this channel", "Choose another channel")
+        if fandomSearch["status"] == "Success" and not getOther:
+            sConfirm = await searchConfirm(ctx, bot, fandomSearch["name"], searchMsg, f"{askTerm} {fandomSearch['name']}?", f"{askTerm} this channel", "Choose another channel")
             if sConfirm["success"]:
                 channelID = await FandomScrape.getChannelURL(fandomSearch["name"])
-                channelID["name"] = fandomSearch["name"]
-                return channelID
+                return {
+                    "success": True,
+                    "channelID": channelID["channelID"],
+                    "name": fandomSearch['name']
+                }
             elif not sConfirm["success"] and not sConfirm["declined"]:
                 await searchMsg.delete()
                 await msgDelete(ctx)
@@ -154,7 +182,7 @@ async def vtuberSearch(ctx: Union[commands.Context, SlashContext], bot: commands
                 }
         
         if not getChannel or fandomSearch["status"] == "Cannot Match":
-            sPick = await searchPrompt(ctx, bot, fandomSearch["results"], searchMsg, "Select a channel to subscribe to:")
+            sPick = await searchPrompt(ctx, bot, fandomSearch["results"], searchMsg, f"Select a channel to {askTerm.lower()}:")
             if not sPick["success"]:
                 await searchMsg.delete()
                 await msgDelete(ctx)
@@ -162,8 +190,11 @@ async def vtuberSearch(ctx: Union[commands.Context, SlashContext], bot: commands
                     "success": False
                 }
             channelID = await FandomScrape.getChannelURL(sPick["name"])
-            channelID["name"] = sPick["name"]
-            return channelID
+            return {
+                "success": True,
+                "channelID": channelID["channelID"],
+                "name": sPick['name']
+            }
 
 async def embedContinue(ctx: Union[commands.Context, SlashContext], bot: commands.Bot, embedMsg: discord.Message, section: str, text: str, name: str):
     textLines = text.split("\n")
@@ -280,4 +311,3 @@ async def uplThumbnail(channelID, videoID, live=True):
             logging.error(upload.value)
         else:
             return upload.value
-
